+++ examples/query.rs
@@ -1,0 +1,70 @@
+/**
+ * MIT License
+ *
+ * ssh2-config - Copyright (c) 2021 Christian Visintin
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+use dirs::home_dir;
+use ssh2_config::SshConfig;
+use std::env::args;
+use std::fs::File;
+use std::io::BufReader;
+use std::path::{Path, PathBuf};
+use std::process::exit;
+
+fn main() {
+    // get args
+    let args: Vec<String> = args().collect();
+    let address = match args.get(1) {
+        Some(addr) => addr.to_string(),
+        None => {
+            usage();
+            exit(255)
+        }
+    };
+    // check path
+    let config_path = match args.get(2) {
+        Some(p) => PathBuf::from(p),
+        None => {
+            let mut p = home_dir().expect("Failed to get home_dir for guest OS");
+            p.extend(Path::new(".ssh/config"));
+            p
+        }
+    };
+    // Open config file
+    let config = read_config(config_path.as_path());
+    let params = config.query(address.as_str());
+    println!("Configuration for {}: {:?}", address, params);
+}
+
+fn usage() {
+    eprintln!("Usage: cargo run --example query -- <address> [config-path]");
+}
+
+fn read_config(p: &Path) -> SshConfig {
+    let mut reader = match File::open(p) {
+        Ok(f) => BufReader::new(f),
+        Err(err) => panic!("Could not open file '{}': {}", p.display(), err),
+    };
+    match SshConfig::default().parse(&mut reader) {
+        Ok(config) => config,
+        Err(err) => panic!("Failed to parse configuration: {}", err),
+    }
+}