+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Changelog
 
 - [Changelog](#changelog)
+  - [0.1.2](#012)
   - [0.1.1](#011)
   - [0.1.0](#010)
 
 ---
+
+## 0.1.2
+
+Released on 11/01/2022
+
+- Implemented `IgnoreUnknown` parameter
+- Added `UseKeychain` support for MacOS
 
 ## 0.1.1